+++ packages/abstract-document/CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file. The format 
 - Add feature to allow setting border colors of cells separately
 - Fix cell border widths not setting properly
 - Add table cell parameter "rowSpan" to allow setting the number of rows a cell should span
+- Allow tables to be split and continue on the next page if they overflow a page
 
 ### Removed